// src/types/project.ts
export interface Project {
  slug: string;
  content: string;
  title?: string;
  image?: string;
  description?: string;
  challenges?: string[];
  descrption?: string;
  featured?: boolean;
  github?: string;
  isPublished?: boolean;
  learnings?: string[];
  live?: string;
  role?: string;
  status: string;
  team?: string;
  teachnologies?: string[];
  timeline?: string;
  ytLink?: string;
}